feat(agents): introduce built-in agent loader with structured definitions
- Added a new `loader.rs` file to manage built-in agent definitions, allowing for easy addition of new agents through a structured format.
- Implemented the `load_builtins` function to parse agent definitions from TOML files and inject system prompts.
- Updated `mod.rs` to include the new loader module, streamlining the organization of agent-related code.
- Ensured that the built-in agents are loaded in a stable order and validated against their defined IDs for consistency.
